refactor: remove any instance or usage of scss
