+++ src/routes/(app)/+error.svelte
@@ -8,17 +8,3 @@
         <h1>Page not found. Take your medicine and try again.</h1>
     </center>
 </div>
-
-
-<style lang="scss">
-    h1 {
-        font-size: 2rem;
-    }
-    img {
-        max-width: 60%;
-    }
-
-    div {
-        padding: 1em 0px;
-    }
-</style>
