+++ src/routes/(app)/Footer.svelte
@@ -2,23 +2,3 @@
     <!-- TODO: Add in buttons here. -->
     <p>&copy; 2025 <a href="/">Michael Sean Brian B. Omisol</a>.</p>
 </footer>
-
-<style lang="scss">
-    footer {
-        text-align: center;
-        
-        p {
-            color: #C0C0C0;
-            font-size: 12px;
-        }
-        a {
-            font-weight: normal;
-            color: #C0C0C0;
-            font-size: 12px;
-
-            &:hover {
-                text-decoration: underline;
-            }
-        }
-    }
-</style>
