+++ src/routes/(app)/about/+page.svelte
@@ -89,108 +89,3 @@
 
 
 </div>
-
-<style lang="scss">
-
-
-    p {
-        text-align: justify;
-    }
-
-    .subtitle {
-        font-style: italic;
-        text-align: center;
-    }
-
-    .other-block {
-        margin: 3em 0 1.2em 1.2em;
-    }
-
-    #authorimg {
-        float: right;
-        margin: 0 1.2em 1.2em 1.2em;
-
-        img {
-            text-align: center;
-        }
-        figcaption {
-            margin:auto;
-            text-align: center;
-            font-size: 0.6rem;
-        }
-    }
-
-    .flex-container {
-        display: flex;
-        flex-wrap: wrap;
-        gap: 50px; 
-    }
-
-    /*
-        https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_grid_layout/Basic_concepts_of_grid_layout#grid_lines
-    */
-    .skills-tab {
-       
-        flex: 1 1 0%;
-        min-width: 220px;
-        max-width: 250px;
-        margin: auto;
-        margin-top: 0;
-        justify-content: center;
-        
-        h2 {
-            height: 2rem;
-            font-size: 1rem;
-            text-align: center;
-            margin-bottom: 1rem;
-        }
-
-
-
-        .images-part {
-            display: grid;
-            grid-template-columns: repeat(4, 1fr);
-            grid-auto-rows: minmax(100px, auto);
-            
-            align-items: center;
-            justify-items: center;
-
-            img {
-                display: inline-grid;
-                width: 50px;
-                filter: drop-shadow(0px 0px 2px black);
-            }
-        }
-    }
-
-
-    .forced-top {
-        margin-top: 20px;
-
-        h1 {
-            margin-bottom: -0.5em;
-        }
-    }
-
-    .section {
-        width: 100%;
-        p {
-            font-size: 0.9rem;
-        }
-        h1 {
-            text-align: center;
-        }
-    }
-
-    @media only screen and (max-width: 537px){
-        #authorimg {
-            float: none;
-            margin: 0 0 -2em 0;
-            text-align: center;
-        }
-        .other-block {
-            margin: auto;
-            margin-top: 3em;
-        }
-    }
-</style>
