+++ src/routes/(app)/contact/+page.svelte
@@ -25,62 +25,3 @@
         </div>
     </div>
 </div>
-
-<style lang="scss">
-
-    
-    // h1 {
-    //     display: block;
-    //     margin-bottom: 0;
-    // }
-
-    p {
-        text-align: center;
-    }
-
-    // #authorimg {
-    //     float: right;
-    //     margin: 0 0 1.2em 2.4em;
-
-    //     figcaption {
-    //         margin:auto;
-    //         text-align: center;
-    //         font-size: 0.6rem;
-    //     }
-    // }
-
-    .section {
-        display: flex;
-        height: 85vh;
-        //padding: 14px;
-        
-        flex-direction: column;
-        justify-content: center;
-        text-align: center;
-    }
-
-    .border-section {
-        position: relative;
-        min-width: 228px;
-        max-width: 300px;
-        padding: 40px 20px;
-        border: 1px solid $border-color;   
-    }
-
-    .images-part {
-        display: grid;
-        grid-template-columns: repeat(3, 50px);
-        grid-auto-rows: auto;
-        
-        justify-content: space-evenly;
-        justify-items: center;
-        align-content: space-evenly;
-        align-items: center;
-        column-gap: 0.5em;
-
-        img {
-            display: inline-grid;
-            width: 50px;
-        }
-    }
-</style>
