+++ src/routes/(app)/projects/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-    import ProjectItem from "$lib/ProjectItem.svelte";
+    import ProjectItem from "./ProjectItem.svelte";
 </script>
 
 <head>
@@ -38,98 +38,3 @@
         <ProjectItem name="Multithreaded GREP Runner" date="2023" link="https://github.com/FlamingHerb/Multithreaded-GREP-Runner"  image="mtgr" />
     </div>
 </div>
-
-<style lang="scss">
-
-    .mini-container {
-        display: flex;
-        align-items: stretch;
-        justify-content: center;
-        flex-wrap: wrap;
-        
-        //overflow: hidden;
-        //background-color: #ffffff;
-        width: 100%;
-        margin: auto;
-
-        column-gap: 20px;
-    }
-
-    :global(figure) {
-        flex: 1;
-        margin: 0.5em 0;
-        min-width: 223.2px;
-        max-width: 223.2px;
-    }
-
-    :global(figcaption) {
-        margin-top: 0.5em;
-        text-align: center;
-        font-size: 1em;
-
-    }
-
-    :global(.poster) {
-        display: block;
-        margin: auto;
-        width: 100%;
-    }
-
-
-    @media only screen and (max-width: $global-max-width){
-        :global(figure) {
-            min-width: 150px;
-            max-width: 150px;
-        }
-        :global(figcaption) {
-            font-size: 0.75em;
-        }
-    }
-
-    h2 {
-        font-family: "Wittgenstein";
-        font-weight: bold;
-        text-align: center;
-        color: $border-color;
-        margin-bottom: 0;
-        font-size: 1.875rem;
-    }
-
-    // p {
-    //     text-align:justify;
-    // }
-
-    // table {
-    //     border-collapse: collapse;
-    //     max-width: 80%;
-        
-    //     tr:nth-child(even) {background-color: #f2f2f2;}
-
-    //     h1 {
-    //         margin-top: 0.25em;
-    //         font-size: 0.75rem;
-
-    //         a {
-    //             text-decoration: none;
-    //         }
-    //     }
-    //     p {
-    //         font-size: 0.5rem;
-    //     }
-
-    //     .description{
-    //         vertical-align: top;
-    //         padding-left: 0.5em;
-    //         padding-right: 0.5em;
-    //     }
-    // }
-
-    /* START:   Mobile responsive area! */
-
-    
-
-
-    /* END:     Mobile responsive area! */
-    
-
-</style>
