+++ src/routes/(app)/projects/ProjectItem.svelte
