+++ src/routes/langleyteaser/+page.svelte
@@ -70,148 +70,3 @@
 <center>
     <img class="clipboard" src="/images/langley/clip_bottom.png" alt="top of clipboard">
 </center>
-
-<style lang="scss">
-    @font-face {
-        font-family: 'DFLiSong Std W7';
-        font-style: normal;
-        font-weight: 400;
-        src: url('/fonts/DFLiSongStd-W7.otf') format('opentype');
-    }
-
-    @font-face {
-        font-family: 'SUIT SemiBold';
-        font-style: normal;
-        font-weight: 400;
-        src: url('/fonts/SUIT-SemiBold.ttf') format('truetype');
-    }
-
-    @font-face {
-        font-family: 'HYQiHeiB5-55';
-        font-style: normal;
-        font-weight: 400;
-        src: url('/fonts/HYQiHeiB5-55.ttf') format('truetype');
-    }
-
-    $breakpoint-tablet: 768px;
-    
-    @media (max-width: $breakpoint-tablet) {
-        div.div1 {
-            width: 100% !important;
-            padding: 0 !important;
-            
-            margin-top: -5px !important;
-            margin-bottom: -10px !important;
-        }
-
-        p {
-            font-size: 0.6rem !important;
-        }
-
-        .ps2 {
-            text-indent: 1em !important;
-        }
-
-        #border {
-            width: 90% !important;
-        }
-
-        .clipboard {
-            width: 100% !important;
-        }
-
-        #text-contain {
-            margin: auto;
-            width: 80%;
-        }
-    }
-
-    .clipboard {
-        width: 850px;
-    }
-
-    :global(body) {
-        padding: 2em;
-        background-color: #000 !important;
-        color: #000;
-    }
-
-    .ps3 {
-        text-align: center;
-    }
-
-    .ps2 {
-        text-indent: 1.5em;
-        text-align: justify;
-        
-        font-family: "HYQiHeiB5-55";
-    }
-
-    .ps1 {
-        text-align: center;
-    }
-
-    div.div1 {
-        background-origin: border-box;
-        background-image: url("/images/langley/clip_background.png");
-        background-size: 100% 100%;
-        background-repeat: no-repeat;
-        background-position: center;
-        width: 650px;
-        padding: 0px 100px;
-        margin: 0 auto;
-        
-        margin-top: -4px;
-        margin-bottom: -16px;
-    }
-
-    span.s1 {
-        color: #7f7f7f;
-        font-size: 1.00rem
-    }
-
-    span.Scrivener-converted-space {
-        white-space: pre;
-    }
-
-    #warning {
-        font-family: "SUIT SemiBold";
-        
-        margin-top: -140px;
-        text-align: center;
-        
-        h1 {
-            font-size: 40px;
-            font-family: "DFLiSong Std W7";
-            letter-spacing: 15px;
-            color: #C91533;
-        }
-
-        #fadeout {
-            margin-top: -60px;
-            padding-top: 40px;
-
-            background-color: #000;
-            background-image: linear-gradient(180deg, #C9DCEF, #000);
-            -webkit-background-clip: text;
-            -moz-background-clip: text;
-            -webkit-text-fill-color: transparent; 
-            -moz-text-fill-color: transparent;
-
-            padding-bottom: 20px;
-            margin-bottom: -20px;
-        }
-
-        
-    }
-
-    #border {
-        margin: auto;
-        margin-top: 48px;
-        margin-bottom: 48px;
-        background-image: url("/images/langley/warning_line.png");
-        background-repeat: repeat-x;
-        height: 12px;
-        width:850px;
-    }
-</style>
